+++ NEWS.md
@@ -11,6 +11,7 @@
 - Improved `get_bb(..., format_out = "sf_polygon")` to return full metadata
   along with geometries (#338 thanks to @RegularnaMatrica)
 - Mention key-only feature requests in README (#342 thanks to @joostschouppe)
+- Merge any columns in `osmdata_sf()` with mixed-case duplicated names (#348)
 
 
 0.2.5